Replace ValueGenerator type class with record type
The upside of this is a clearer model and the use of less advanced constructs.
Both of which lead to clearer code in a variety of places. Oh, and simpler,
more concise type signatures.

We loose the ability to have a non-generic show instance for ValueGenerators.

I took the suggestions for the refactoring from a question on stackoverflow[1]
with the title "How to avoid typing the same context over and over again?"

[1]: http://stackoverflow.com/questions/6706248