Experiment more closely matches the real program
And now clearly demonstrates the problem I'm having:

A type [Metric a], can contain elements that are either Metric Double or it
could contain elements that are Metric String, but it can't contain both.

The introduction of a new type, MetricType, say, may solve this by allowing me
to replace [Metric a] with [Metric MetricType]. With a different value
constructor for the different supported metric types.