Hosts can have multiple type safe metrics
Instead of having a type

  data Metric a = Metric { metricValue_    :: MetricValue a
                         , metricValueGen_ :: ValueGenerator a
                         }

and trying to create a type safe hetrogeneous list of metrics, and still be
able to run the ValueGenerator runGenerator function, we have the types

  data MetricRecord a = Metric { metricValue_ :: MetricValue a }
  data Metric = forall a. MkM (MetricRecord a) (MetricRecord a -> Metric)

Metric can be seen to be type safe. Constructing lists of Metric is simple,
but what can be done with the elements of the list is limited.

It is possible to run the function contained in the MetricTup to get back a
new MetricTup, it is also possible to print the value of the MetricValue which
is several levels below MetricTup.

Hopefully these will prove sufficient to implement hmond proper.