Update dependencies to build with GHC 7.6.1
I've removed the upper bounds on the libraries shipped with GHC. My reasoning
is that 1) every time a new GHC is released, any new versions shipped with GHC
will be used; having an upper bound results in the dependency resolution
failing, 2) as these dependencies are shipped with GHC, anyone trying to
compile this project will have versions of these packages already installed
and Cabal will pick those rather than installing new versions.

This leaves us with the situation that that the new version of a GHC-shipped
dependency is not backwards compatible. In that case the either the compile
will fail or there will be a runtime error. In either case, this is what we
would have got even with upper bounds, only updating the upper bounds is
skipped.